rename: layout -> components
